bloc3-pipelines-donnees/fraud-detection-pipeline/requirements.txt
# ─── Core Pipeline ─────────────────────────────────────────────
apache-airflow==2.7.3
psycopg2-binary==2.9.9
requests==2.31.0
pandas==2.1.4
numpy==1.26.2

# ─── ML Pipeline ──────────────────────────────────────────────
scikit-learn==1.3.2
joblib==1.3.2
mlflow==2.9.2

# ─── Data Quality ─────────────────────────────────────────────
great-expectations==0.18.8

# ─── Testing ──────────────────────────────────────────────────
pytest==7.4.3
